rename client -> RTC
