+++ packages/react-client/example/src/App.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import { TelnyxClientProvider } from '@telnyx/react-client';
+import { TelnyxRTCProvider } from '@telnyx/react-client';
 import ClientStatus from './ClientStatus';
 import CallLog from './CallLog';
 
@@ -12,10 +12,10 @@ const App = () => {
   };
 
   return (
-    <TelnyxClientProvider credential={credential}>
+    <TelnyxRTCProvider credential={credential}>
       <ClientStatus />
       <CallLog />
-    </TelnyxClientProvider>
+    </TelnyxRTCProvider>
   );
 };
