--- packages/react-client/src/TelnyxClientContext.tsx
@@ -1,6 +1,0 @@
-import React from 'react';
-import { TelnyxRTC } from '@telnyx/webrtc';
-
-const TelnyxClientContext = React.createContext<TelnyxRTC | null>(null);
-
-export default TelnyxClientContext;
