+++ packages/react-client/src/TelnyxRTCContext.tsx
@@ -1,0 +1,6 @@
+import React from 'react';
+import { TelnyxRTC } from '@telnyx/webrtc';
+
+const TelnyxRTCContext = React.createContext<TelnyxRTC | null>(null);
+
+export default TelnyxRTCContext;
