+++ packages/react-client/src/index.tsx
@@ -1,5 +1,5 @@
-export { default as TelnyxClientContext } from './TelnyxClientContext';
-export { default as TelnyxClientProvider } from './TelnyxClientProvider';
-export { default as useTelnyxClient } from './useTelnyxClient';
+export { default as TelnyxRTCContext } from './TelnyxRTCContext';
+export { default as TelnyxRTCProvider } from './TelnyxRTCProvider';
+export { default as useTelnyxRTC } from './useTelnyxRTC';
 export { default as useEvents } from './useEvents';
 export { default as useNotification } from './useNotification';
